Сделал

+++ MPIpr/MPIpr/Process.h
@@ -6,22 +6,22 @@
 class Process {
 private:
     int rank; //ранг процесса
-    long value; // значение
-    long recVal; // временное хранилизе для полученого значения от прошлого процесса
+    long long value; // значение
+    long long recVal; // временное хранилизе для полученого значения от прошлого процесса
     int step; // щаг
-    long rec2bVal;
+    long long rec2bVal;
 
 public:
-    Process(int rank, long value);
+    Process(int rank, long long value);
     void send();
     void recv();
 
     // методы для доступа к данным
-    long getValue() const;
-    void setValue(long newValue);
-    void setRecVal(long newRecVal);
-    long getRecVal() const;
-    long getRec2bVal() const;
+    long long getValue() const;
+    void setValue(long long newValue);
+    void setRecVal(long long newRecVal);
+    long long getRecVal() const;
+    long long getRec2bVal() const;
 };
 
 #endif
